fix(setup): close the four live findings from the #733 review
Seventeen of the twenty-one open review threads were already closed by
"fix(setup): close GH719 recovery gaps"; each was re-checked against the
current tree. The four that still held:

- setup_managed_tree_remove.rs: an active quarantine's stored source prefix
  is historical evidence install-state preflight already validated. Forcing
  the request's new prefix aborted every install and re-enable after a
  manifest moved a disabled skill's source directory, with setup having
  already refreshed the installed snapshot. Ownership is now proven against
  the recorded prefix, so the move is idempotent.
- install-state.sh: state_prepare_clean treated a directory or symlink at an
  install-state path as absent, so --clean removed managed Claude/Codex assets
  and only then failed on the malformed path. Both paths are now rejected
  before any removal.
- install-state.sh: state_check_drift inspected only the current generation,
  so a previous-only quarantine left by an interrupted release reported CLEAN
  while the next install would abort on it.
- codex-home.sh: a disabled skill path that is not an exact install-state-owned
  copy was reported as a neutral [DISABLED] promising removal, but the
  quarantine installer refuses that tree. It is now [BROKEN], which --strict
  surfaces as exit 2.

The source-prefix change contradicted active_quarantine_rejects_a_manifest_
source_move, which asserted the hard reject. Per maintainer decision the
contract is now idempotent pass-through; the test was rewritten to assert the
existing quarantine is returned and that neither the record nor the retained
bytes are re-attributed.

setup_managed_tree_remove.rs would have crossed the U-16 ceiling, so its
durable-IO and path helpers moved to setup_managed_tree_io.rs.
