+++ vibeguard-runtime/src/setup_managed_tree_io.rs
@@ -1,0 +1,153 @@
+//! Durable filesystem primitives and small value predicates shared by the
+//! managed-tree quarantine lifecycle. Split out of
+//! `setup_managed_tree_remove.rs` to keep that file under the U-16 ceiling.
+
+use crate::setup_support::{SetupResult, write_json_atomic};
+use serde_json::Value;
+use std::ffi::CString;
+use std::fs::{self, File, OpenOptions};
+use std::io::{self, Write};
+use std::path::{Path, PathBuf};
+use std::time::{SystemTime, UNIX_EPOCH};
+
+pub(super) fn write_new_json_durable(path: &Path, value: &Value) -> SetupResult<()> {
+    let tmp = path.with_extension(format!("tmp.{}.{}", std::process::id(), now_nanos()));
+    let bytes = serde_json::to_vec_pretty(value)?;
+    let result = (|| {
+        let mut file = OpenOptions::new().write(true).create_new(true).open(&tmp)?;
+        file.write_all(&bytes)?;
+        file.write_all(b"\n")?;
+        file.sync_all()?;
+        rename_noreplace(&tmp, path)?;
+        sync_parent(path)
+    })();
+    if result.is_err() {
+        let _ = fs::remove_file(&tmp);
+    }
+    result
+}
+
+pub(super) fn write_json_durable(path: &Path, value: &Value) -> SetupResult<()> {
+    write_json_atomic(path, value)?;
+    sync_parent(path)
+}
+
+pub(super) fn sync_parent(path: &Path) -> SetupResult<()> {
+    sync_directory(path.parent().ok_or("durable path has no parent")?)
+}
+
+pub(crate) fn sync_directory(path: &Path) -> SetupResult<()> {
+    #[cfg(unix)]
+    File::open(path)?.sync_all()?;
+    #[cfg(windows)]
+    let _ = path;
+    Ok(())
+}
+
+pub(super) fn ensure_public_absent(dest: &Path) -> SetupResult<()> {
+    match fs::symlink_metadata(dest) {
+        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
+        Ok(_) => Err(format!(
+            "disabled public destination unexpectedly exists: {}",
+            dest.display()
+        )
+        .into()),
+        Err(error) => Err(error.into()),
+    }
+}
+
+pub(super) fn valid_text(value: &Value) -> bool {
+    value
+        .as_str()
+        .is_some_and(|text| !text.is_empty() && !text.contains(['\n', '\r']))
+}
+
+pub(super) fn valid_digest(value: &Value) -> bool {
+    value
+        .as_str()
+        .and_then(|text| text.strip_prefix("sha256:"))
+        .is_some_and(|digest| {
+            digest.len() == 64
+                && digest
+                    .bytes()
+                    .all(|byte| byte.is_ascii_hexdigit() && !byte.is_ascii_uppercase())
+        })
+}
+
+#[cfg(any(target_os = "linux", target_os = "android"))]
+pub(crate) fn rename_noreplace(from: &Path, to: &Path) -> io::Result<()> {
+    use std::os::unix::ffi::OsStrExt;
+    let from = CString::new(from.as_os_str().as_bytes())?;
+    let to = CString::new(to.as_os_str().as_bytes())?;
+    let result = unsafe {
+        libc::renameat2(
+            libc::AT_FDCWD,
+            from.as_ptr(),
+            libc::AT_FDCWD,
+            to.as_ptr(),
+            libc::RENAME_NOREPLACE,
+        )
+    };
+    if result == 0 {
+        Ok(())
+    } else {
+        Err(io::Error::last_os_error())
+    }
+}
+
+#[cfg(any(target_os = "macos", target_os = "ios"))]
+pub(crate) fn rename_noreplace(from: &Path, to: &Path) -> io::Result<()> {
+    use std::os::unix::ffi::OsStrExt;
+    let from = CString::new(from.as_os_str().as_bytes())?;
+    let to = CString::new(to.as_os_str().as_bytes())?;
+    let result = unsafe { libc::renamex_np(from.as_ptr(), to.as_ptr(), libc::RENAME_EXCL) };
+    if result == 0 {
+        Ok(())
+    } else {
+        Err(io::Error::last_os_error())
+    }
+}
+
+#[cfg(windows)]
+pub(crate) fn rename_noreplace(from: &Path, to: &Path) -> io::Result<()> {
+    match fs::rename(from, to) {
+        Ok(()) => Ok(()),
+        Err(_) if fs::symlink_metadata(to).is_ok() => Err(io::ErrorKind::AlreadyExists.into()),
+        Err(error) => Err(error),
+    }
+}
+
+#[cfg(not(any(
+    target_os = "linux",
+    target_os = "android",
+    target_os = "macos",
+    target_os = "ios",
+    windows
+)))]
+pub(crate) fn rename_noreplace(_from: &Path, _to: &Path) -> io::Result<()> {
+    Err(io::Error::new(
+        io::ErrorKind::Unsupported,
+        "atomic no-replace rename is unsupported",
+    ))
+}
+
+pub(super) fn absolute(path: &Path) -> PathBuf {
+    if path.is_absolute() {
+        path.to_path_buf()
+    } else {
+        std::env::current_dir()
+            .unwrap_or_else(|_| PathBuf::from("."))
+            .join(path)
+    }
+}
+
+pub(super) fn path_text(path: &Path) -> String {
+    path.to_string_lossy().into_owned()
+}
+
+pub(crate) fn now_nanos() -> u128 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_nanos()
+}
